% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.models.R
\name{combine.models}
\alias{combine.models}
\title{Combine suitable habitat models based on geographic criteria}
\usage{
combine.models(
  mod1,
  mod2,
  y.breaks = NULL,
  x.breaks = NULL,
  buffer.width = 15000,
  drop.crumbs = 1000,
  res = 500
)
}
\arguments{
\item{mod1, mod2}{...}

\item{y.breaks}{Latitude break for the model combination as a character vector including logical operators. For example \code{c("<2.5e6", ">=2.5e6")}}

\item{x.breaks}{Not implemented.}

\item{buffer.width}{Single numeric value defining the \link[rgeos]{gBuffer} parameter for the smoothing of polygon output.}

\item{drop.crumbs}{Single numeric value specifying a threshold (area in km2) for small disconnected polygons which should be removed from the polygonized version of the suitable habitat. Set to 0 (or \code{NA}) to bypass the removal. Uses the \link[smoothr]{drop_crumbs} function.}

\item{res}{Vertical and horizontal resolution of the smoothed polygon output.}

\item{proj4}{Character argument specifying the \link[sp]{proj4string} (projection) for the smoothed polygon output.}
}
\description{
Used to combine (currently only two) suitable habitat models based on (currently linear) geographic criteria
}
