% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat.space.R
\name{habitat.space}
\alias{habitat.space}
\title{Estimate habitat space using multiple constraining variables}
\usage{
habitat.space(
  data = NULL,
  cvars,
  log.transform = rep(FALSE, length(cvars)),
  rvar = NULL,
  grid.size = rep(30, length(cvars)),
  grid.limits = NULL,
  method = "kde.binary",
  non.suitable.level = 0.001,
  chull.correction = TRUE
)
}
\arguments{
\item{data}{data.frame containing the data for habitat space estimation. Required except when all limits are set orthogonal.}

\item{cvars}{A named list defining the constraining variables for the habitat space. Each name refers to a variable name in \code{data} unless orthogonal limits are desired. In this case, the name refers to the variable name and the element must be a numeric vector of length two defining the minimum and maximum tolerance, respectively. If this is not the case, the variables are assumed simultaneously limiting and acquired from \code{data}. The variables will be plotted in the order given here with the first variable being understood as the x-axis, the second as the y-axis, the third as the z-axis, and so on. At the moment up to three list elements are allowed. This limit may be increased in future releases.}

\item{log.transform}{A logical vector with the same length than \code{cvars} indicating whether the variables should be 10-based logarithm transformed before habitat space estimation.}

\item{rvar}{a character argument specifying the column name for the response variable if not the same as the constraining variables (\code{cvars}). Set to \code{NULL} (default) to use the density of constraining variables as a response variable (i.e. presence-only data).}

\item{grid.size}{a numeric vector equally long as the number of simultaneously constraining variables (dimensions; \code{cvars}) defining the grid.size for binning which is required to estimate habitat spaces.}

\item{grid.limits}{a list of numeric vectors, each of length two defining the minimum and maximum limit for constraining variables (\code{cvars}). The list should be the same length length than the number of simultaneously constraining variables. Set to \code{NULL} to automatically estimate the limits.}

\item{method}{a character argument specifying the method for probability estimation for the habitat space. Currently implemented methods:
\itemize{
\item \code{"kde.binary"} Multivariate kernel density estimation using a binary probability level to define suitable and not-suitable habitat in the multivariate space. The \code{non.suitable.level} argument sets the probability level.
\item \code{"kde.probability"} Multivariate kernel density estimation is used to calculate the probability for finding the species in a given region in the habitat space.
\item \code{"foo"} Frequencies of occurrence for each bin are used as probabilities for finding the species in a given region in the habitat space. Requires a response variable (\code{rvar}). If the response variable is not binary (1 or 0 / \code{TRUE} or \code{FALSE}, it will be converted to such.
\item \code{"quantile.regression"} Multivariate quantile regression. Not implemented yet. 
\item function of glm or gam family Multivariate glm or gam object. Not implemented yet. 

Note that setting all limits to orthogonal will lead to a rectangular habitat space and probabilities will not be calculated.
}}

\item{non.suitable.level}{a numeric argument defining the KDE probability or frequency of occurrence level for non-suitable habitat. The value can be set between 0 and 1. Defines the edge of the habitat space. Smaller the value, the larger the habitat space.}

\item{chull.correction}{logical indicating whether convex hull (\link{chull}) should be applied to correct concave shapes along the edge of the habitat space.}
}
\description{
The function estimates a multivariate habitat space for a species. The results of this function can be used in further plotting and suitable habitat estimation functions in the package.
}
\details{
The function retunrs only two dimensional habitat spaces at the moment.
}
\author{
Mikko Vihtakari
}
