% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdiffcoords.R
\name{hdiffcoords}
\alias{hdiffcoords}
\title{Print overlapping and unique bin centroid coordinates for two hexbin objects}
\usage{
hdiffcoords(bin1, bin2)
}
\arguments{
\item{bin1, bin2}{two objects of class hexbin.}
}
\value{
Returns a list of data frames with unique coordinates for \code{bin1} and \code{bin2} as well as overlapping coordinates among bins.
}
\description{
Print overlapping and unique bin centroid coordinates for two hexbin objects
}
\details{
The hexbin objects for comparison, bin1 and bin2, must have the same plotting limits and cell size.
}
\keyword{internal}
