% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexagonize.suitable.habitat.R
\name{hexagonize.suitable.habitat}
\alias{hexagonize.suitable.habitat}
\title{Hexagonize suitable habitat}
\usage{
hexagonize.suitable.habitat(data, hexbins)
}
\arguments{
\item{data}{a raster object from the \code{\link{rasterize.suitable.habitat}} function.}

\item{hexbins}{A number of (xbins) used for the hexagonization. See \code{\link[hexbin]{hexbin}}.}
}
\description{
An internal function to hexagonize suitable habitat data from the \code{\link{suitable.habitat}} function
}
\keyword{internal}
