% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting.factors.R
\name{limiting.factors}
\alias{limiting.factors}
\title{Define factors limiting suitable habitat}
\usage{
limiting.factors(habitat.space, var.cols, dt, svars.dt)
}
\arguments{
\item{habitat.space}{An object from the \code{\link{habitat.space}} function.}

\item{var.cols}{Character vector from the \code{\link{suitable.habitat}} function.}

\item{dt}{Data frame from the \code{\link{suitable.habitat}} function.}

\item{svars.dt}{Data frame from the \code{\link{suitable.habitat}} function.}
}
\description{
An internal function to calculate factors limiting suitable habitat used in the \code{\link{suitable.habitat}} function
}
\keyword{internal}
