% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.fit.R
\name{model.fit}
\alias{model.fit}
\title{Examine fit of observations to a suitable habitat model}
\usage{
model.fit(mod, pts, regions = NULL, method = "hexagon")
}
\arguments{
\item{mod}{A \link[=suitable.habitat]{suitable habitat model object}.}

\item{pts}{\link[sp]{SpatialPoints} object containing observations of the species. If \code{proj4string} is missing, it is assumed to be equal to \code{mod}.}

\item{regions}{}

\item{method}{hexagon, raster, polygon}
}
\value{
Returns a list containing suitable habitat estimates in following formats: 1) \strong{\code{$raw}:} oceanographic model (\code{oceangr.model}) grid cells, 2) \strong{\code{$raster}:} rasterized cells whose resolution is defined by the \code{res} with disconnected regions smaller than the \code{drop.crumbs} parameter removed (in km2), 3) \strong{\code{$polygon}:} polgyonized raster data, and 4) \strong{\code{$hexagon}:} hexagonized raster data. See \code{\link[hexbin]{hexbin}}.
}
\description{
Compares a suitable habitat model to observations specified as spatial points.
}
\details{
...
}
