% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{NEMOdata}
\alias{NEMOdata}
\title{Oceanographic NEMO model pan-Arctic data for the 00s}
\format{NEMOdata}
\usage{
NEMOdata
}
\description{
Oceanographic NEMO model pan-Arctic data for the 00s
}
\keyword{datasets}
