% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.habitatSpace.R
\name{plot.habitatSpace}
\alias{plot.habitatSpace}
\title{Plot a habitat space}
\usage{
\method{plot}{habitatSpace}(x, ...)
}
\arguments{
\item{x}{\code{habitatSpace} object from \code{\link{habitat.space}} function.}

\item{...}{Additional graphical arguments. Required by R build checks. Ignore.}
}
\description{
Plot method for \code{\link[=habitat.space]{habitatSpace}} objects.
}
\seealso{
\code{\link{habitat.space}}
}
\author{
Mikko Vihtakari
}
