% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.NEMOdata.R
\name{plot.NEMOdata}
\alias{plot.NEMOdata}
\title{Plot a NEMO data on a map}
\usage{
\method{plot}{NEMOdata}(x, limits = 40, type = "all", midpoint = NA, cutpoints = NA, ...)
}
\arguments{
\item{x}{\code{NEMOdata} object}

\item{limits}{Map limits. See \code{\link{basemap}}. The option "auto" limits the map using coordinate range of \code{x}. Alternatively use a numeric vector as described in \code{\link[PlotSvalbard]{basemap}} documentation.}

\item{type}{Character argument defining the types of model data to be plotted: 
\itemize{
\item \code{"depth"} Depth data as raster
\item \code{"temp"} Temperature data as raster
\item \code{"sal"} Salinity data as raster
\item \code{"all"} All of the above
}}

\item{...}{Additional arguments. Required by R build checks. Ignore.}
}
\description{
Plot method for \code{\link[=NEMOdata]{NEMOdata}} objects.
}
\author{
Mikko Vihtakari
}
