% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-docs.R
\docType{data}
\name{polarStereographic}
\alias{polarStereographic}
\title{Character string containing the CRS projection used by the NEMO model}
\format{character}
\usage{
polarStereographic
}
\description{
Character string containing the CRS projection used by the NEMO model
}
\keyword{datasets}
