% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.habitatSpace.R
\name{print.habitatSpace}
\alias{print.habitatSpace}
\title{Print habitat space (\code{habitatSpace}) objects}
\usage{
\method{print}{habitatSpace}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=habitat.space]{habitatSpace}} objects
}
\seealso{
\code{\link{basemap}}
}
\author{
Mikko Vihtakari
}
\keyword{internal}
