% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.NEMOdata.R
\name{print.NEMOdata}
\alias{print.NEMOdata}
\title{Print NEMO data (\code{NEMOdata}) objects}
\usage{
\method{print}{NEMOdata}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=suitable.habitat]{NEMOdata}} objects
}
\author{
Mikko Vihtakari
}
\keyword{internal}
