% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SHfit.R
\name{print.SHfit}
\alias{print.SHfit}
\title{Print suitable habitat model fit (\code{SHfit}) objects}
\usage{
\method{print}{SHfit}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=model.fit]{SHfit}} objects
}
\author{
Mikko Vihtakari
}
\keyword{internal}
