% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.SHmod.R
\name{print.SHmod}
\alias{print.SHmod}
\title{Print habitat space (\code{SHmod}) objects}
\usage{
\method{print}{SHmod}(x, ...)
}
\arguments{
\item{x}{object to be printed.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
\code{\link{print}} function for \code{\link[=suitable.habitat]{SHmod}} objects
}
\author{
Mikko Vihtakari
}
\keyword{internal}
