% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.nonsuitable.habitat.R
\name{rasterize.nonsuitable.habitat}
\alias{rasterize.nonsuitable.habitat}
\title{Rasterize non-suitable habitat to find limiting factors}
\usage{
rasterize.nonsuitable.habitat(
  data,
  proj4,
  mod.extent = mod.ext,
  drop.crumbs,
  res
)
}
\arguments{
\item{data}{a data frame from \code{\link{suitable.habitat}} function}

\item{proj4}{Character argument specifying the \link[sp]{proj4string} (projection) for the smoothed polygon output.}
}
\description{
An internal function to rasterize non-suitable habitat data from the \code{\link{suitable.habitat}} function
}
\seealso{
\code{\link{limiting.factors}} \code{\link{plot.SHmod}}
}
\keyword{internal}
