% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize.suitable.habitat.R
\name{rasterize.suitable.habitat}
\alias{rasterize.suitable.habitat}
\title{Rasterize suitable habitat}
\usage{
rasterize.suitable.habitat(data, proj4, drop.crumbs, res)
}
\arguments{
\item{data}{a data frame from \code{\link{suitable.habitat}} function}

\item{proj4}{Character argument specifying the \link[sp]{proj4string} (projection) for the smoothed polygon output.}

\item{drop.crumbs}{Single numeric value specifying a threshold (area in km2) for small disconnected polygons which should be removed from the polygonized version of the suitable habitat. Set to 0 (or \code{NA}) to bypass the removal. Uses the \link[smoothr]{drop_crumbs} function.}

\item{res}{Vertical and horizontal resolution of the smoothed polygon output.}
}
\description{
An internal function to rasterize suitable habitat data from the \code{\link{suitable.habitat}} function
}
\keyword{internal}
