% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suitable.habitat.R
\name{suitable.habitat}
\alias{suitable.habitat}
\title{Estimate a suitable habitat for a species from NEMO output}
\usage{
suitable.habitat(
  habitat.space,
  oceangr.model = NEMOdata,
  buffer.width = 15000,
  proj4 = polarStereographic,
  lat.lim = 40,
  res = 400,
  find.lim.factors = FALSE,
  drop.crumbs = 1000
)
}
\arguments{
\item{habitat.space}{...}

\item{oceangr.model}{...}

\item{buffer.width}{Single numeric value defining the \link[rgeos]{gBuffer} parameter for the smoothing of polygon output.}

\item{proj4}{Character argument specifying the \link[sp]{proj4string} (projection) for the smoothed polygon output.}

\item{lat.lim}{Single numeric value defining a latitude value which is used to cut the model data.}

\item{res}{Vertical and horizontal resolution of the smoothed polygon output.}

\item{find.lim.factors}{Logical indicating whether the function should which factors (if any) make the NEMO grid cell non-suitable habitat.}

\item{drop.crumbs}{Single numeric value specifying a threshold (area in km2) for small disconnected polygons which should be removed from the polygonized version of the suitable habitat. Set to 0 (or \code{NA}) to bypass the removal. Uses the \link[smoothr]{drop_crumbs} function.}
}
\description{
Estimates suitable habitat for a species from European Modeling of the Ocean (NEMO) physical oceanography model output based on limitations set by temperature, salinity and bottom depth.
}
\details{
The \code{\link[base]{dim}}ensions of NEMO input matrices must be equal.
}
