% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.models.R
\name{combine.models}
\alias{combine.models}
\title{Combine suitable habitat models based on geographic criteria}
\usage{
combine.models(
  mod1,
  mod2,
  y.breaks = NULL,
  x.breaks = NULL,
  buffer.width = 15000,
  drop.crumbs = 30000,
  res = 350,
  hexbins = 100
)
}
\arguments{
\item{mod1, mod2}{\link[=suitable.habitat]{Suitable habitat models} to be be combined.}

\item{x.breaks, y.breaks}{Longitude and/or latitude breaks for the model combination as a character vector of length 2 including logical operators. The corresponding y/x intercept will be used to combine the model parts. Should be given as the model \code{\link[=suitable.habitat]{proj4}} units. For example \code{c("<2.5e6", ">=2.5e6")}}

\item{buffer.width}{Single numeric value defining the \link[rgeos]{gBuffer} parameter for the smoothing of polygon output.}

\item{drop.crumbs}{Single numeric value specifying a threshold (area in km2) for small disconnected polygons which should be removed from the polygonized version of the suitable habitat. Set to 0 (or \code{NA}) to bypass the removal. Uses the \link[smoothr]{drop_crumbs} function.}

\item{res}{Vertical and horizontal resolution of the smoothed polygon output.}

\item{hexbins}{A number of (xbins) used for the hexagonization. See \code{\link[hexbin]{hexbin}}.}
}
\description{
Used to combine (currently only two) suitable habitat models based on (currently linear) geographic criteria
}
\details{
The function finds limiting factors if both \code{mod1} and \code{mod2} contain them.
}
