% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SHfit.R
\name{plot.SHfit}
\alias{plot.SHfit}
\title{Plot a suitable habitat model on a map}
\usage{
\method{plot}{SHfit}(
  x,
  limits = "auto",
  type = x$parameters$fit.method,
  region.plot = x$parameters$regions,
  axis.labels = TRUE,
  title = "capitalize",
  base_size = 8,
  ...
)
}
\arguments{
\item{x}{\code{SHfit} object from the \code{\link{model.fit}} function.}

\item{limits}{Map limits. See \code{\link{basemap}}. The option "auto" (default) limits the map using coordinate range of \code{mod}. Alternatively use a numeric vector as described in \code{\link[PlotSvalbard]{basemap}} documentation.}

\item{type}{Character argument defining the types of model data to be plotted: 
\itemize{
\item \code{"polygon"} Model fit based on polygons and spatial points,
\item \code{"hexagon"} Hexagon griding based model fit.
\item \code{"raster"} Model fit based on raster grid.
}}

\item{title}{"capitalize", NA, anything else}

\item{...}{Additional arguments to \code{\link[PlotSvalbard]{basemap}}}
}
\description{
Plot method for \code{\link[=model.fit]{SHfit}} objects.
}
\seealso{
\code{\link{suitable.habitat}} \code{\link{habitat.space}}
}
\author{
Mikko Vihtakari
}
