% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SHmod.R
\name{plot.SHmod}
\alias{plot.SHmod}
\title{Plot a suitable habitat model on a map}
\usage{
\method{plot}{SHmod}(x, limits = "auto", ...)
}
\arguments{
\item{x}{\code{SHmod} object from \code{\link{suitable.habitat}} function.}

\item{limits}{Map limits. See \code{\link{basemap}}. The option "auto" (default) limits the map using coordinate range of \code{mod}. Alternatively use a numeric vector as described in \code{\link[PlotSvalbard]{basemap}} documentation.}

\item{...}{Additional arguments. Required by R build checks. Ignore.}
}
\description{
Plot method for \code{\link[=suitable.habitat]{SHmod}} objects.
}
\seealso{
\code{\link{suitable.habitat}} \code{\link{habitat.space}}
}
\author{
Mikko Vihtakari
}
