% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SHmod.R
\name{plot.SHmod}
\alias{plot.SHmod}
\title{Plot a suitable habitat model on a map}
\usage{
\method{plot}{SHmod}(x, limits = "auto", type = "default", legend.position = "right", ...)
}
\arguments{
\item{x}{\code{SHmod} object from \code{\link{suitable.habitat}} function.}

\item{limits}{Map limits. See \code{\link{basemap}}. The option "auto" (default) limits the map using coordinate range of \code{mod}. Alternatively use a numeric vector as described in \code{\link[PlotSvalbard]{basemap}} documentation.}

\item{type}{Character argument defining the types of model data to be plotted: 
\itemize{
\item \code{"raw"} Only raw data. I.e. the grid cell data directly from the oceanographic model.
\item \code{"raster"} Only raster data. See \code{\link{suitable.habitat}} and \code{\link{rasterize.suitable.habitat}}
\item \code{"polygon"} Only polygon data. See \code{\link{suitable.habitat}} and \code{\link{polygonize.suitable.habitat}}
\item \code{"hexagon"} Only hexagon data. See \code{\link{suitable.habitat}} and \code{\link{hexagonize.suitable.habitat}}
\item \code{"raw&raster"} Options \code{"raw"} and \code{"raster"} on top of each other.
\item \code{"lim.fact"} Limiting factors as raster data.
\item \code{"default"} Options \code{"raw"}, \code{"raster"}, and \code{"polygon"} on top of each other.
}}

\item{...}{Additional arguments. Required by R build checks. Ignore.}
}
\description{
Plot method for \code{\link[=suitable.habitat]{SHmod}} objects.
}
\seealso{
\code{\link{suitable.habitat}} \code{\link{habitat.space}}
}
\author{
Mikko Vihtakari
}
